+++ src/terraform_manager/terraform/versions.py
@@ -1,12 +1,9 @@
 import os
 import textwrap
-from typing import List, Dict, Union, Optional
+from typing import List, Dict
 
-from requests import Response
 from tabulate import tabulate
-from terraform_manager.entities.error_response import ErrorResponse
 from terraform_manager.entities.workspace import Workspace
-from terraform_manager.terraform.workspaces import batch_operation
 
 VersionSummary = Dict[str, List[Workspace]]
 
@@ -45,90 +42,6 @@ def check_versions(workspaces: List[Workspace], new_version: str) -> bool:
         if workspace.is_terraform_version_newer_than(new_version):
             return False
     return True
-
-
-def patch_versions(
-    terraform_domain: str,
-    organization: str,
-    workspaces: List[Workspace],
-    *,
-    new_version: str,
-    no_tls: bool = False,
-    token: Optional[str] = None,
-    write_output: bool = False
-) -> bool:
-    """
-    Patches the Terraform version of the workspaces.
-
-    :param terraform_domain: The domain corresponding to the targeted Terraform installation (either
-                             Terraform Cloud or Enterprise).
-    :param organization: The organization containing the workspaces to patch.
-    :param workspaces: The workspaces to patch.
-    :param new_version: The new Terraform version to assign to the workspaces.
-    :param no_tls: Whether to use SSL/TLS encryption when communicating with the Terraform API.
-    :param token: A token suitable for authenticating against the Terraform API. If not specified, a
-                  token will be searched for in the documented locations.
-    :param write_output: Whether to print a tabulated result of the patch operations to STDOUT.
-    :return: Whether all patch operations were successful. If even a single one failed, returns
-             False.
-    """
-
-    if not check_versions(workspaces, new_version):
-        return False
-
-    json = {"data": {"type": "workspaces", "attributes": {"terraform-version": new_version}}}
-    report = []
-
-    def on_success(workspace: Workspace) -> None:
-        if workspace.terraform_version == new_version:
-            report.append([
-                workspace.name,
-                workspace.terraform_version,
-                new_version,
-                "success",
-                "version unchanged"
-            ])
-        else:
-            report.append([
-                workspace.name, workspace.terraform_version, new_version, "success", "none"
-            ])
-
-    def on_failure(workspace: Workspace, response: Union[Response, ErrorResponse]) -> None:
-        report.append([
-            workspace.name,
-            workspace.terraform_version,
-            workspace.terraform_version,
-            "error",
-            response.json()
-        ])
-
-    result = batch_operation(
-        terraform_domain,
-        workspaces,
-        json=json,
-        on_success=on_success,
-        on_failure=on_failure,
-        no_tls=no_tls,
-        token=token,
-        write_output=write_output
-    )
-
-    if write_output:
-        print((
-            f'Terraform workspace version patch results for organization "{organization}" at '
-            f'"{terraform_domain}":'
-        ))
-        print()
-        print(
-            tabulate(
-                sorted(report, key=lambda x: (x[3], x[0])),
-                headers=["Workspace", "Version Before", "Version After", "Status", "Message"],
-                colalign=("left", "right", "right")
-            )
-        )
-        print()
-
-    return result
 
 
 def write_version_summary(