+++ src/terraform_manager/terraform/workspaces.py
@@ -1,24 +1,36 @@
 import itertools
 from fnmatch import fnmatch
-from typing import List, Optional, Dict, Any, Callable, Union
+from typing import List, Optional, Dict, Any, Callable, Union, TypeVar
 
 import requests
 from requests import Response
+from tabulate import tabulate
 from terraform_manager.entities.error_response import ErrorResponse
 from terraform_manager.entities.workspace import Workspace
 from terraform_manager.terraform import pagination, get_api_headers, SuccessHandler, ErrorHandler
 from terraform_manager.utilities.throttle import throttle
 from terraform_manager.utilities.utilities import safe_http_request, get_protocol
 
+A = TypeVar("A")
+
 
 def _map_workspaces(json: List[Dict[str, Any]]) -> List[Workspace]:
+    required_attributes = [
+        "name", "terraform-version", "auto-apply", "locked", "working-directory", "execution-mode"
+    ]
+
+    def is_valid(json_object: Dict[str, Any]) -> bool:
+        if "id" not in json_object or "attributes" not in json_object:
+            return False
+        else:
+            for attribute in required_attributes:
+                if attribute not in json_object["attributes"]:
+                    return False
+            return True
+
     workspaces = []
     for json_object in json:
-        if "id" in json_object and "attributes" in json_object and "name" in \
-                json_object["attributes"] and "terraform-version" in json_object["attributes"] and \
-                "auto-apply" in json_object["attributes"] and "locked" in \
-                json_object["attributes"] and "working-directory" in json_object["attributes"] and \
-                "execution-mode" in json_object["attributes"]:
+        if is_valid(json_object):
             workspaces.append(
                 Workspace(
                     json_object["id"],
@@ -82,7 +94,7 @@ def fetch_all(
     ]
 
 
-def batch_operation(
+def _internal_batch_operation(
     terraform_domain: str,
     workspaces: List[Workspace],
     *,
@@ -93,26 +105,6 @@ def batch_operation(
     token: Optional[str] = None,
     write_output: bool = False
 ) -> bool:
-    """
-    Patches workspaces in batch fashion using a given JSON body, executing callback functions
-    depending on the success or failure of each individual PATCH operation.
-
-    :param terraform_domain: The domain corresponding to the targeted Terraform installation (either
-                             Terraform Cloud or Enterprise).
-    :param workspaces: The workspaces to patch.
-    :param json: The JSON body of the PATCH requests that will be sent to the Terraform API.
-    :param on_success: A function which will be passed a Workspace object when that workspace has
-                       been successfully patched.
-    :param on_failure: A function which will be passed a Workspace object when that workspace has
-                       NOT been successfully patched.
-    :param no_tls: Whether to use SSL/TLS encryption when communicating with the Terraform API.
-    :param token: A token suitable for authenticating against the Terraform API. If not specified, a
-                  token will be searched for in the documented locations.
-    :param write_output: Whether to print a tabulated result of the patch operations to STDOUT.
-    :return: Whether all patch operations were successful. If even a single one failed, returns
-             False.
-    """
-
     headers = get_api_headers(terraform_domain, token=token, write_error_messages=write_output)
     all_successful = True
     base_url = f"{get_protocol(no_tls)}://{terraform_domain}/api/v2"
@@ -127,3 +119,94 @@ def batch_operation(
             all_successful = False
             on_failure(workspace, response)
     return all_successful
+
+
+def batch_operation(
+    terraform_domain: str,
+    organization: str,
+    workspaces: List[Workspace],
+    *,
+    field_mapper: Callable[[Workspace], A],
+    field_name: str,
+    new_value: A,
+    report_only_value_mapper: Optional[Callable[[A], str]] = None,
+    no_tls: bool = False,
+    token: Optional[str] = None,
+    write_output: bool = False
+) -> bool:
+    """
+    Patches workspaces in batch fashion using a given JSON body, executing callback functions
+    depending on the success or failure of each individual PATCH operation.
+
+    :param terraform_domain: The domain corresponding to the targeted Terraform installation (either
+                             Terraform Cloud or Enterprise).
+    :param organization: The organization containing the workspaces to patch.
+    :param workspaces: The workspaces to patch.
+    :param field_mapper: A function which will be passed a Workspace object that must return a field
+                         value to be compared against the new_value and printed in the tabulated
+                         report.
+    :param field_name: The Terraform API-compatible name of the field returned by field_mapper (e.g.
+                       "auto-apply" instead of "auto apply" or "Auto Apply").
+    :param new_value: The value to update the workspace field to (where the field corresponds to
+                      field_name).
+    :param report_only_value_mapper: A function which will be passed the result of the field_mapper
+                                     function when writing the tabulated report's "Before" and
+                                     "After" columns. If not specified, the default will be the
+                                     str() function.
+    :param no_tls: Whether to use SSL/TLS encryption when communicating with the Terraform API.
+    :param token: A token suitable for authenticating against the Terraform API. If not specified, a
+                  token will be searched for in the documented locations.
+    :param write_output: Whether to print a tabulated result of the patch operations to STDOUT.
+    :return: Whether all patch operations were successful. If even a single one failed, returns
+             False.
+    """
+
+    json = {"data": {"type": "workspaces", "attributes": {field_name: new_value}}}
+    report = []
+
+    report_mapper = str if report_only_value_mapper is None else report_only_value_mapper
+
+    def on_success(workspace: Workspace) -> None:
+        report.append([
+            workspace.name,
+            report_mapper(field_mapper(workspace)),
+            report_mapper(new_value),
+            "success",
+            f"{field_name} unchanged" if field_mapper(workspace) == new_value else "none"
+        ])
+
+    def on_failure(workspace: Workspace, response: Union[Response, ErrorResponse]) -> None:
+        report.append([
+            workspace.name,
+            report_mapper(field_mapper(workspace)),
+            report_mapper(field_mapper(workspace)),
+            "error",
+            response.json()
+        ])
+
+    result = _internal_batch_operation(
+        terraform_domain,
+        workspaces,
+        json=json,
+        on_success=on_success,
+        on_failure=on_failure,
+        no_tls=no_tls,
+        token=token,
+        write_output=write_output
+    )
+
+    if write_output:
+        print((
+            f'Terraform workspace {field_name} patch results for organization "{organization}" '
+            f'at "{terraform_domain}":'
+        ))
+        print()
+        print(
+            tabulate(
+                sorted(report, key=lambda x: (x[3], x[0])),
+                headers=["Workspace", "Before", "After", "Status", "Message"]
+            )
+        )
+        print()
+
+    return result