+++ tests/terraform/test_versions.py
@@ -2,14 +2,13 @@ import os
 from typing import List
 from unittest.mock import MagicMock, call
 
-import responses
 from pytest_mock import MockerFixture
 from tabulate import tabulate
 from terraform_manager.entities.workspace import Workspace
 from terraform_manager.terraform.versions import group_by_version, write_version_summary, \
-    VersionSummary, check_versions, patch_versions
+    VersionSummary, check_versions
 
-from tests.utilities.tooling import test_workspace, TEST_API_URL, TEST_TERRAFORM_DOMAIN, \
+from tests.utilities.tooling import test_workspace, TEST_TERRAFORM_DOMAIN, \
     TEST_ORGANIZATION
 
 _0_12_28: Workspace = test_workspace(version="0.12.28")
@@ -45,63 +44,6 @@ def test_check_versions() -> None:
     assert not check_versions(_workspaces, "0.13.0")
     assert check_versions(_workspaces, "0.13.5")
     assert check_versions(_workspaces, "0.13.8")
-
-
-@responses.activate
-def test_patch_versions(mocker: MockerFixture) -> None:
-    for test in [_0_13_1_first.terraform_version, "0.13.5"]:
-        mocker.patch("terraform_manager.terraform.credentials.find_token", return_value="test")
-        print_mock: MagicMock = mocker.patch("builtins.print")
-        error_json = {"data": {"id": _0_13_1_second.workspace_id}}
-        responses.add(
-            responses.PATCH, f"{TEST_API_URL}/workspaces/{_0_13_1_first.workspace_id}", status=200
-        )
-        responses.add(
-            responses.PATCH,
-            f"{TEST_API_URL}/workspaces/{_0_13_1_second.workspace_id}",
-            json=error_json,
-            status=500
-        )
-        assert not patch_versions(
-            TEST_TERRAFORM_DOMAIN,
-            TEST_ORGANIZATION, [_0_13_1_first, _0_13_1_second],
-            new_version=test,
-            write_output=True
-        )
-        # yapf: disable
-        print_mock.assert_has_calls([
-            call((
-                f'Terraform workspace version patch results for organization "{TEST_ORGANIZATION}" '
-                f'at "{TEST_TERRAFORM_DOMAIN}":'
-            )),
-            call(),
-            call(
-                tabulate(
-                    [
-                        [
-                            _0_13_1_second.name,
-                            _0_13_1_second.terraform_version,
-                            _0_13_1_second.terraform_version,
-                            "error",
-                            str(error_json)
-                        ],
-                        [
-                            _0_13_1_first.name,
-                            _0_13_1_first.terraform_version,
-                            test,
-                            "success",
-                            "none" if test != _0_13_1_first.terraform_version
-                            else "version unchanged"
-                        ]
-                    ],
-                    headers=["Workspace", "Version Before", "Version After", "Status", "Message"],
-                    colalign=("left", "right", "right"))
-
-            ),
-            call()
-        ])
-        # yapf: enable
-        assert print_mock.call_count == 4
 
 
 def test_write_version_summary(mocker: MockerFixture) -> None: